将代码进行了小幅度整合
