+++ app/build.gradle
@@ -11,7 +11,7 @@ android {
     }
 
     compileSdkVersion 26
-    buildToolsVersion "26.0.1"
+    buildToolsVersion '26.0.2'
     defaultConfig {
         applicationId "com.example.myapplication"
         minSdkVersion 16
