+++ app/src/main/res/layout/activity_login.xml
@@ -52,7 +52,6 @@
         android:layout_width="match_parent"
         android:layout_height="20dp"
         android:gravity="right"
-        android:text="忘记密码?"
         android:layout_marginTop="12dp" />
     <LinearLayout
         android:layout_width="300dp"
